Update geometry field in the initial migration
The geometry field is referenced in the views which prevent us alter the field in migrations.
Since we are not really changing the schema, we can apply the correct field definition the initial
migration.

Previously it uses the default (wrong) srid, which does not match the database srid and django will prevent saving objects. Also we change the field name from geometry_1 to geometry while preserving the db_column.